add an example for memory
