--- docs/docs/Memory/example.md
@@ -1,4 +1,0 @@
----
-id: Example
-sidebar_position: 2
----
